+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Here we provide notes that summarize the most important changes in each released
 
 Please consult the changelog to inform yourself about breaking changes and security issues.
 
+## [v0.2.1](https://github.com/Materials-Data-Science-and-Informatics/fair-python-cookiecutter/tree/v0.2.1) <small>(2024-02-13)</small> { id="0.2.1" }
+
+* fixed a bug where instantiation of template fails if no repository URL is provided
+* improve usability and description of some CLI fields
+
 ## [v0.2.0](https://github.com/Materials-Data-Science-and-Informatics/fair-python-cookiecutter/tree/v0.2.0) <small>(2024-01-05)</small> { id="0.2.0" }
 
 * Refactored into custom tool `fair-python-cookiecutter` for more versatility and convenience